requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.5
frozenlist==1.4.1
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.19
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==1.0.3
langchain-text-splitters==0.0.1
langsmith==0.1.56
markdown-it-py==3.0.0
marshmallow==3.21.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.3
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.1.19
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
rich==13.7.1
rsa==4.9
SQLAlchemy==2.0.30
tenacity==8.3.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
